[nvim] Improve structure of plugins directory
