+++ config/nvim/lua/steve/keymaps.lua
@@ -54,6 +54,6 @@ map("v", "/", "/\\v", "Use search in visual mode", { silent = false })
 
 -- Search and replace (substitute) prompt
 -- stylua: ignore start
-map("n", "gs", ":%s/\\v", "Search and replace (vim substitute)", { silent = false })
-map("v", "gs", ":s/\\v", "Search and replace (vim substitute)", { silent = false })
-map("n", "gS", ":%s/\\<<C-r><C-w>\\>//g<Left><Left>", "Search and replace word under cursor (vim substitute)", { silent = false })
+map("n", "gs", ":%s/\\v", "Go: substitute", { silent = false })
+map("v", "gs", ":s/\\v", "Go: subsitute", { silent = false })
+map("n", "gS", ":%s/\\<<C-r><C-w>\\>//g<Left><Left>", "Go: substitute", { silent = false })
