+++ config/nvim/lua/steve/plugins/editor_nvim_spectre.lua
@@ -1,0 +1,24 @@
+-- Search/Replace in Multiple Files
+-- nvim-spectre provides powerful search and replace capabilities across multiple files.
+return {
+  {
+    "nvim-pack/nvim-spectre",
+    config = function()
+      require("spectre").setup({
+        highlight = {
+          replace = "@text.warning",
+        },
+      })
+    end,
+    -- stylua: ignore
+    keys = {
+      -- Search and Replace all files in current directory
+      { "<leader>sr", '<cmd>lua require("spectre").open()<CR>',                               desc = "Open Search and Replace (Spectre)" },
+      { "<leader>sR", '<cmd>lua require("spectre").open_visual({select_word=true})<CR>',      desc = "Search and Replace word under cursor (Spectre)" },
+      { "<leader>sR", '<esc><cmd>lua require("spectre").open_visual()<CR>', mode = "x",       desc = "Search and Replace current selection (Spectre)" },
+      -- Search and Replace on current file
+      { "<leader>sb", '<cmd>lua require("spectre").open_file_search()<CR>',                   desc = "Search and Replace on current file (Spectre)" },
+      { "<leader>sB", '<cmd>lua require("spectre").open_file_search({select_word=true})<CR>', desc = "Search and Replace word under cursor on current file (Spectre)" },
+    },
+  },
+}
