+++ config/nvim/lua/steve/plugins/ide_copilot.lua
@@ -1,0 +1,7 @@
+-- copilot
+return {
+  {
+    "github/copilot.vim",
+    event = "VeryLazy",
+  },
+}
